move element displays to separate files
