+++ lib/forked/display/display.rb
@@ -1,5 +1,9 @@
-require_relative 'display_paragraph.rb'
 require_relative 'display_blockquote.rb'
+require_relative 'display_button.rb'
+require_relative 'display_code_block.rb'
+require_relative 'display_heading.rb'
+require_relative 'display_paragraph.rb'
+require_relative 'display_rule.rb'
 
 $gtk.reset
 
@@ -363,245 +367,6 @@ module Forked
       end
     end
 
-    #######################
-    # DISPLAY ELEMENT TYPES
-    #######################
-
-    ### BUTTON
-    # TODO: prevent buttons from being instantiated if object already exists
-    # TODO: wrap up button generation code and put it in the button module
-    def display_button(y_pos, item, _content, __i)
-      button = data.style.button
-      display = data.style.display
-      button_box = data.style.button_box
-
-      button.size_px = args.gtk.calcstringbox('X', button.size_enum, button.font)[1]
-
-      # previous element is also a button? use spacing_between instead of spacing_after
-      if @last_printed_element_type == :button
-        y_pos += button.spacing_after * button.size_px
-        y_pos -= button.spacing_between * button.size_px
-      end
-
-      if item.action.empty?
-        button = data.style.disabled_button
-        button_box = data.style.disabled_button_box
-      end
-
- 
-      text_w, button.size_px = args.gtk.calcstringbox(item.text, button.size_enum, button.font)
-      text_w = text_w.to_i
-      button_h = (button.size_px + button_box.padding_top + button_box.padding_bottom)
-
-      x = display.margin_left
-      y = (y_pos - button_h).to_i
-      w = text_w + button_box.padding_left + button_box.padding_right
-      h = (button.size_px + button_box.padding_top + button_box.padding_bottom).to_i
-
-      pill_button_base = { x: x, y: y, w: w, h: h, text: item.text, action: item.action }
-
-      enabled = Effed.pill_button_layer(
-        args,
-        rect: pill_button_base,
-        color: { r: button.r, g: button.g, b: button.b, a: button.a || 255 },
-        bg: { r: button_box.r, g: button_box.g, b: button_box.b, a: button.a || 255 },
-        font: button.font,
-        size_enum: button.size_enum
-      )
-      focused = Effed.pill_button_layer(
-        args,
-        rect: pill_button_base,
-        color: {
-          r: data.style.selected_button.r,
-          g: data.style.selected_button.g,
-          b: data.style.selected_button.b,
-          a: data.style.selected_button.a || 255
-        },
-        bg: {
-          r: data.style.selected_button_box.r,
-          g: data.style.selected_button_box.g,
-          b: data.style.selected_button_box.b,
-          a: data.style.selected_button_box.a || 255
-        },
-        font: data.style.selected_button.font,
-        size_enum: data.style.selected_button.size_enum
-      )
-      active = Effed.pill_button_layer(
-        args,
-        rect: pill_button_base,
-        color: {
-          r: data.style.active_button.r,
-          g: data.style.active_button.g,
-          b: data.style.active_button.b,
-          a: data.style.active_button.a || 255
-        },
-        bg: {
-          r: data.style.active_button_box.r,
-          g: data.style.active_button_box.g,
-          b: data.style.active_button_box.b,
-          a: data.style.active_button.a || 255
-        },
-        font: data.style.active_button.font,
-        size_enum: data.style.active_button.size_enum
-      )
-      disabled = Effed.pill_button_layer(
-        args,
-        rect: pill_button_base,
-        color: {
-          r: data.style.disabled_button.r,
-          g: data.style.disabled_button.g,
-          b: data.style.disabled_button.b,
-          a: data.style.disabled_button.a || 255
-        },
-        bg: {
-          r: data.style.disabled_button_box.r,
-          g: data.style.disabled_button_box.g,
-          b: data.style.disabled_button_box.b,
-          a: data.style.disabled_button_box.a || 255
-        },
-        font: data.style.disabled_button.font,
-        size_enum: data.style.disabled_button.size_enum
-      )
-
-      option = Effed::FButton.new(
-        rect: pill_button_base,
-        enabled: enabled,
-        focused: focused,
-        active: active,
-        disabled: disabled
-      )
-
-      option.force_status_change(:disabled) if item.action.empty?
-
-      data.options << option if item&.action && !item.action.empty?
-      data.primitives << option
-
-      y_pos - (button_box.padding_top +
-               button_box.padding_bottom +
-               button.size_px +
-               button.size_px * button.spacing_after)
-    end
-
-    ### HEADING
-
-    def display_heading(y_pos, item, _previous_element_type)
-      heading = data.style.heading
-      display = data.style.display
-      heading.size_px = args.gtk.calcstringbox('X', heading.size_enum, heading.font)[1]
-
-      if heading.align
-        canvas_w = display.w
-        x_offset = canvas_w * heading.align
-        x_pos = display.margin_left + x_offset
-        anchor_x = heading.align
-      end
-
-      data.primitives << {
-        x: x_pos || display.margin_left,
-        y: y_pos.to_i,
-        text: item.text,
-        anchor_x: anchor_x || 0
-      }.label!(heading)
-
-      y_pos - heading.size_px * heading.spacing_after
-    end
-
-    ### RULE
-
-    def display_rule(y_pos, item, _previous_element_type)
-      rule = data.style.rule
-      display = data.style.display
-      weight = rule.weight
-      weight = item.weight if item.weight
-
-      data.primitives << {
-        x: display.margin_left,
-        y: y_pos.to_i,
-        w: display.w,
-        h: weight
-      }.sprite!(rule)
-
-      y_pos - rule.spacing_after
-    end
-
-    def display_image(y_pos, item)
-      image = data.style.image
-      display = data.style.display
-
-      # I don't know why path doesn't work *as is*
-      # It's a string, and it doesn't have any odd
-      # characters in it
-      # But putting it into a _new_ string makes it work
-      path = item[:path].to_s
-      merger = {}
-      if path.end_with?('}') && path.include?('{')
-        path, remnant = Forked::Parser.pull_out('{', '}', path)
-        merger = eval('{' + remnant + '}')
-      end
-
-      h = merger.h || 80
-      # w = merger.w || 80
-
-      im = {
-        x: display.margin_left,
-        y: y_pos.to_i - h,
-        w: 80,
-        h: h,
-        path: path
-      }.sprite!(image)
-
-      data.primitives << im.merge(merger)
-
-      y_pos - (h + image.spacing_after)
-    end
-
-    ### CODE BLOCK
-
-    def display_code_block(y_pos, item, _previous_element_type)
-      code_block = data.style.code_block
-      display = data.style.display
-      code_block_box = data.style.code_block_box
-
-      text_array = wrap_lines_code_block(
-        item.text, code_block.font, code_block.size_enum,
-        display.w - (code_block_box.padding_left + code_block_box.padding_right)
-      )
-      code_block.size_px = args.gtk.calcstringbox('X', code_block.size_enum, code_block.font)[1]
-
-      box_height = text_array.count * (code_block.size_px * code_block.line_spacing) +
-                   code_block_box.padding_top + code_block_box.padding_bottom
-
-      temp_y_pos = y_pos
-
-      rect = {
-        x: display.margin_left, 
-        y: y_pos - box_height.to_i, 
-        w: display.w, h: box_height,
-        r: code_block_box.r,
-        g: code_block_box.g,
-        b: code_block_box.b
-      }
-      bg = Effed.rounded_panel(rect: rect)
-      data.primitives << bg
-
-      temp_y_pos -= code_block_box.padding_top
-      data.primitives << text_array.map do |line|
-        label = {
-          x: display.margin_left + code_block_box.padding_left,
-          y: temp_y_pos.to_i,
-          text: line
-        }.label!(code_block)
-
-        temp_y_pos -= code_block.size_px * code_block.line_spacing
-
-        label
-      end
-
-      y_pos -= box_height
-      # y_pos - code_block.size_px * code_block.spacing_after
-      y_pos -= code_block_box.margin_bottom
-    end
-
     ################
     # STYLE HANDLING
     ################
@@ -621,12 +386,6 @@ module Forked
         data.style.paragraph
       end
     end
-
-    ############
-    # PRIMITIVES
-    ############
-
-
 
     ###############
     # TEXT HANDLING
